requirements.txt
requests>=2.31.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
google-generativeai>=0.3.0
pillow>=10.0.0
ipython>=8.0.0
polygon-api-client>=2.0.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0 
pytickertick>=0.1.1